fix bug

+++ src/main/scala/com/cpc/spark/common/CpcPartitioner.scala
@@ -7,7 +7,7 @@ import scala.util.hashing.MurmurHash3
 class CpcPartitioner(num: Int) extends Partitioner {
   override def numPartitions: Long = num.toLong
 
-  override def getPartition(key: Any): Int = {
+  override def getPartition(key: Any): Long = {
     val k = key.asInstanceOf[(String, Int)]
     (Murmur3Hash.stringHash(k._1) % numPartitions + numPartitions + k._2) % numPartitions
   }
